chore: resolve some comments
