+++ cookbook/05_agent_os/factories/team/01_basic_team_factory.py
@@ -15,7 +15,7 @@ Test:
 """
 
 from agno.agent import Agent
-from agno.agent.factory import RequestContext
+from agno.factory import RequestContext
 from agno.db.postgres import PostgresDb
 from agno.models.openai import OpenAIResponses
 from agno.os import AgentOS
